feat: Refactor AdminDashboard structure, update routing, and enhance sidebar functionality
