+++ src/Routing.jsx
@@ -4,7 +4,7 @@ import { Routes, Route } from "react-router-dom";
 import LoginPage from "./pages/LoginPage";
 import ClientRegistration from "./pages/Agent-newUser";
 import Home from "./pages/Home";
-import AgentDashboard from "./pages/AgentDashboard/AdminDashboard";
+import AdminWindow from "./components/AdminDashboard/AgentWindow";
 
 const Routing = () => {
   return (
@@ -13,7 +13,7 @@ const Routing = () => {
       <Route path="/login" element={<LoginPage />} />
       <Route path="/" element={<Home />} />
       <Route path="/user" element={<ClientRegistration />} />
-      <Route path="/agent" element={<AgentDashboard />} />
+      <Route path="/agent/*" element={<AdminWindow />} />
     </Routes>
   );
 };
