+++ src/pages/AgentDashboard/Dashboard.jsx
@@ -4,7 +4,7 @@ import React from "react";
 const Dashboard = () => {
   return (
     <div>
-      <h2>Dashboard</h2>
+      <h2>Dashboard </h2>
       {/* Add your content here */}
     </div>
   );
